Farkle 0.0.2
Changelog

-Added functions scoreHand and checkForBust

scoreHand
-checks for combinations according to rules given by the KCD wiki page
https://kingdom-come-deliverance.fandom.com/wiki/Dice
-adds corresponding points to turnPoints

checkForBust
-checks if any scoring combinations are possible given the roll generated
-throws true if no combinations found, false if combinations are possible

Plans/Thought Process

hey guys heres just a short paragraph going through how im thinking these functions can be implemented in the frontend

my thinking right now is to write the majority of the games logic here in farkleBackend, and try to minimize the amount of "cluttered code" which will end up appearing in the frontend of our game. as my plan goes so far the game will start with player 1, running the function initialRoll() to create a set of 6 dice to start the game. from there, we will call the function checkforBust() to see if any combinations are possible in the first place. if none are, then set turnPoints to 0 and then call the function to end and then pass the turn (this is yet to be written). setting turnPoints to 0 will ensure that if a player is re-rolling their dice and they get a bust, then all points they gained previously that turn will be set to 0. however, if combinations are possible, then the player will be allowed the change the currentHandIndex using sensor functionality (yet to be written), and run function selectDie() using sensor functionality (yet to be written). Once they are done selecting (choose method of selecting this option later), run function scoreHand() then the function to add turnPoints to playerScores(currentPlayer) (yet to be written).

after that is all done it seems like most of the backend will be completed for the game, and we can move on to implementing the frontend, animations/sound, and our additional reach goals